requirements.txt
aniso8601==1.3.0
blinker==1.4
certifi==2017.11.5
chardet==3.0.4
click==6.7
Cython==0.27.1
decorator==4.1.2
Flask==0.12.2
Flask-Login==0.4.0
Flask-RESTful==0.3.6
Flask-SQLAlchemy==2.3.2
Flask-WhooshAlchemy==0.56
Flask-WTF==0.14.2
idna==2.6
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
marshmallow==2.15.0
numpy==1.13.3
pandas==0.21.1
pbr==3.1.1
python-dateutil==2.6.1
pytz==2017.3
requests==2.18.4
six==1.11.0
SQLAlchemy==1.1.14
sqlalchemy-migrate==0.11.0
sqlparse==0.2.4
Tempita==0.5.2
urllib3==1.22
Werkzeug==0.12.2
Whoosh==2.7.4
WTForms==2.1
